improvements for item builder with animations
