+++ lib/release_4/controller/reorderable_controller.dart
@@ -6,20 +6,6 @@ abstract class ReorderableController {
   final childrenKeyMap = <dynamic, ReorderableEntity>{};
   final offsetMap = <int, Offset>{};
 
-  void handleCreatedChild({
-    required Offset? offset,
-    required ReorderableEntity reorderableEntity,
-  }) {
-    if (offset != null) {
-      offsetMap[reorderableEntity.updatedOrderId] = offset;
-    }
-    _updateMaps(
-      reorderableEntity: reorderableEntity.creationFinished(
-        offset: offset,
-      ),
-    );
-  }
-
   ReorderableEntity getReorderableEntity({
     required ValueKey key,
     required int index,
@@ -40,19 +26,33 @@ abstract class ReorderableController {
         updatedOffset: offset,
       );
     }
+    // print('get ${reorderableEntity.toString()}');
     return reorderableEntity;
   }
 
-  void handleMovingFinished({required ReorderableEntity reorderableEntity}) {
-    _updateMaps(
-      reorderableEntity: reorderableEntity.positionUpdated(),
-    );
+  void handleCreatedChild({
+    required Offset? offset,
+    required ReorderableEntity reorderableEntity,
+  }) {
+    if (offset != null) {
+      print('created ${reorderableEntity.key} $offset');
+      offsetMap[reorderableEntity.updatedOrderId] = offset;
+    }
+    final updatedEntity = reorderableEntity.creationFinished(offset: offset);
+    // print('creation finished ${updatedEntity.toString()}');
+    _updateMaps(reorderableEntity: updatedEntity);
   }
 
   void handleOpacityFinished({required ReorderableEntity reorderableEntity}) {
-    _updateMaps(
-      reorderableEntity: reorderableEntity.fadedIn(),
-    );
+    final updatedEntity = reorderableEntity.fadedIn();
+    // print('opacity finished ${updatedEntity.toString()}');
+    _updateMaps(reorderableEntity: updatedEntity);
+  }
+
+  void handleMovingFinished({required ReorderableEntity reorderableEntity}) {
+    final updatedEntity = reorderableEntity.positionUpdated();
+    // print('moving finished ${updatedEntity.toString()}');
+    _updateMaps(reorderableEntity: updatedEntity);
   }
 
   void _updateMaps({required ReorderableEntity reorderableEntity}) {
